django-auth/requirements.txt
django
djangorestframework
python-decouple
django-cors-headers
djangorestframework_simplejwt
psycopg2-binary
django-s3-storage